+++ src/pages/matching/Matching.tsx
@@ -1,4 +1,4 @@
-import MatchingContainer from '../components/matching/MatchingContainer';
+import MatchingContainer from '../../components/matching/MatchingContainer';
 
 const MatchingPage = () => {
   return <MatchingContainer></MatchingContainer>;
